Move repro 17212 to nested questions file
